関数の実行完了を検出する。

+++ src/lambchop.h
@@ -37,7 +37,7 @@ bool lambchop_file_read_all(const char *path, lambchop_logger *logger, char **bu
 
 int lambchop_vm_main(void *func, uint64_t stacksize, lambchop_logger *logger);
 int lambchop_vm_run(void *mainfunc, lambchop_logger *logger);
-int lambchop_vm_call(void *func, int argc, uint64_t *argv, lambchop_logger *logger);
+uint64_t lambchop_vm_call(void *func, int argc, uint64_t *argv, lambchop_logger *logger);
 
 uint64_t lambchop_syscall(uint64_t *rax, uint64_t a0, uint64_t a1, uint64_t a2, uint64_t a3, uint64_t a4, uint64_t a5);
